pyproject.toml
[tool.poetry]
name = "qfieldcloud_fetcher"
version = "0.0.1"
description = "A python project to fetch data from EMI QFieldCloud instance and prepare pictures for iNaturalist import."
authors = ["Edouard Brülhart <edouard.bruelhart@unifr.ch>"]
repository = "https://github.com/edouardbruelhart/qfieldcloud-fetcher"
documentation = "https://edouardbruelhart.github.io/qfieldcloud-fetcher/"
readme = "README.md"
packages = [
  {include = "qfieldcloud_fetcher"}
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pytest-cov = "^6.0.0"
python-dotenv = "^1.0.1"
qfieldcloud-sdk = "^0.9.0"
requests = "^2.32.3"
types-requests = "^2.32.0.20241016"
pandas = "^2.2.3"
geopandas = "0.14.4"
pandas-stubs = "2.2.2.240807"
fiona = "^1.10.1"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["qfieldcloud_fetcher"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"



[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true



[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
